Move and restructure the tests for the Contains impl
